Create sequences for Xhibit trial and appeal tables
This commit creates sequences to be used for insertion
into the `XHIBIT_TRIAL_DATA` and `XHIBIT_APPEAL_DATA`
tables. This is done through the `@SequenceGenerator`
and `@GeneratedValues` annotations on the database
entity class, which means that the sequences created
in this commit do not need to be directly hooked-up
to the tables in the database.