Add PrototypeBeanFactory
When calling the XhibitDataService from the TrialDataService,
some instance variables are persisted to the former during
the process of getting data from S3.

By default, Spring Boot creates all beans as singletons, which
means you will get the same bean back every time for a particular-
requested class definition. This is an issue for the XhibitDataService,
given the aforementioned storage of instance variables.

Ideally whenever the TrialDataService invokes the XhibitDataService
in one of its methods, it will get a new instance of the XhibitDataService
to work on, removing any potential race conditions with multiple
consumers otherwise seeking to call methods on the same XhibitDataService.

The `@Scope("prototype")` annotation on the XhibitDataService tells
Spring to create a new instance whenever we request a bean of this class.
At the same time, we must ensure we do not store a `final` version of
said bean anywhere, otherwise we will always access the same bean
which is effectively a singleton.

As part of this commit, a new PrototypeBeanFactory has been created to be
able to mock the XhibitDataService which the framework obtains by overwriting
the `@Lookup` method on the factory. The `@Lookup` annotation is how we tell
Spring to get an instance of the bean from the ApplicationContext, which
then sees it is prototype-scoped and returns a new instance. Without the factory,
it is impossible to mock the `@Lookup` annotated method and subsequently all
tests fail.

Reading:

- https://docs.spring.io/spring-framework/reference/core/beans/dependencies/factory-method-injection.html#beans-factory-lookup-method-injection