Refactor to pass continuation token to data service
This commit refactors the XhibitDataService to move the
continutation token and records retrieved flag out of
instance variables and into the GetRecordSheetsResponse,
removing all state and allowing the XhibitDataService
to return to a singleton instance. This is turn removes
the need for the PrototypeBeanFactory that was introduced
specifically to allow testing and overall simplifies the
approach taken.