Add S3 bucket name back as an environment variable
This commit re-introduces the S3 bucket name as an
environment variables to pods in the cluster. This
was [previously removed](https://github.com/ministryofjustice/laa-maat-scheduled-tasks/pull/28)
as only the dev bucket was created and secret set,
causing a start-up failure for the app in later
environments. Now that buckets are created in all
environments and secrets set accordingly, this
environment variable can be safely re-introduced.