Add missing test for pagination and refactor
This commit adds a missing test for the _XhibitDataService_
`getAllRecordSheets` method. Pagination was previously handled
within the _TrialDataService_, but in a recent commit the
pagination logic was moved back down the stack and into the
_XhibitDataService_. At the same time, tests for the
_TrialDataService_ and _AppealDataService_ assumed what the
_GetRecordSheetsResponse_ would like look but without this
being tested.

This commit also refactors the `buildRecordSheetsResponse`
method in the _XhibitDataService_ to once again be `private`,
as it should only _ever_ be called through the `getAllRecordSheets`
method and therefore does not need exposing.