Add BillingDataFeedLogMapper
This commit adds a new _BillingDataFeedLogMapper_ for mapping from
the _BillingDataFeedLogEntity_ class to the respective DTO (deserialising
the payload within). To achieve this in a way which reduces duplication,
a new base class _BillingDTO_ has been created and is added to the
billing DTO classes. If a _JsonProcessingException_ occurs during
deserialisation then the mapper will return `null` rather than throw an
exception, allowing all correctly-serialised data to be deserialised.