Add controller endpoint and associated tests
This commit adds a new controller endpoint at `/api/internal/v1/billing/resend`
to manually trigger a re-send of billing data to the CCLF endpoints reading
from the database table. To trigger this, the corresponding scheduler service
is updated to add a new method which triggers the appropriate _resend_ methods
on the downstream services.