Implement serialisation and deserialisation using Jackson
This commit fully implements serialisation and
deserialisation using Jackson. Specifically, the
_BillingDataFeedLogMapper_ is updated to handle mapping
any DTO which extends _BillingDTO_ to an entity in a
single method. Methods are added in the same class to map
from an entity to each of the implementations of
_BillingDTO_ to avoid the consumer needing to cast the
return types appropriately. Some duplication arises in
the methods from doing this, but it does make things
easier for the consumer and improves general readability
than having a single generic method which maps from DTO
to entity.