Create BillingService base class for shared billing services shared behaviour
This commit creates a BillingService abstract class to reduce code duplication between RepOrderBillingService, ApplicantBillingService and ApplicantHistoryBillingService.

To achieve this a BillingDTO interface has also been created and is implemented on the ApplicantBillingDTO, ApplicantHistoryBillingDTO and RepOrderBillingDTO allowing this to be used generically across lists in the BillingService code.

Tests and scheduler have also been updated to use the generic method name for sending records to billing.