Refactor BatchProcessingService and downstream services
This commit refactors the _BatchProcessingService_ to
simplify process of invoking the concrete _BillingService_
implementations. It does this by removing the `userModified`
parameter from the methods of those downstream services
and instead obtaining it directly from the _BillingConfiguration_
object which those services already have access to.

The _BatchProcessingService_ is also simplifed by removing
the `TriConsumer` reference and instead referencing
`BillingService<T>` when calling the internal method to
process a batch of _BillingDTO_s.